fix: title and description of play and playoffline routes
